+++ package/thunderstore/Waystones/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+* mouse acceleration in direction search mode
+* gamepad sticks acceleration rebalanced in direction search mode
+* waystone interaction buttons swapped for better gamepad support
+
 # 1.0.1
 * config options to disable directions
 * items from configurable json/yaml list can be sacrificed to waystone to reduce cooldown